+++ slang.h
@@ -194,8 +194,8 @@ extern "C"
         void const*             userData);
 
     /*!
-    @brief Add a path in which source files are being search. When the programmer specifies @code using <file_name> @endcode in code, the compiler searches the file
-    in all search pathes in order.
+    @brief Add a path to use when searching for referenced files.
+    This will be used for both `#include` directives and also for explicit `__import` declarations.
     @param ctx The compilation context.
     @param searchDir The additional search directory.
     */
@@ -203,6 +203,16 @@ extern "C"
         SlangCompileRequest*    request,
         const char*             searchDir);
 
+    /*!
+    @brief Add a path to use when searching for referenced files, that automatically treats `#include` as `__import`
+    This behaves just like `spAddSearchPath()` except that any `#include` file found through this path
+    will be treated as if it was referenced with `__import`.
+    @param ctx The compilation context.
+    @param searchDir The additional search directory.
+    */
+    SLANG_API void spAddAutoImportPath(
+        SlangCompileRequest*    request,
+        const char*             searchDir);
     /*!
     @brief Add a macro definition to be used during preprocessing.
     @param key The name of the macro to define.
@@ -213,6 +223,13 @@ extern "C"
         const char*             key,
         const char*             value);
 
+    /*!
+    @brief Set options using arguments as if specified via command line.
+    */
+    SLANG_API int spProcessCommandLineArguments(
+        SlangCompileRequest*    request,
+        char const* const*      args,
+        int                     argCount);
 
     /** Add a distinct translation unit to the compilation request
 
@@ -890,6 +907,7 @@ namespace slang
 #include "source/slang/diagnostics.cpp"
 #include "source/slang/emit.cpp"
 #include "source/slang/lexer.cpp"
+#include "source/slang/options.cpp"
 #include "source/slang/parameter-binding.cpp"
 #include "source/slang/parser.cpp"
 #include "source/slang/preprocessor.cpp"
